refactor: consolidate presentation logic and extract payload data
