+++ packages/cli/src/reports/markdown.ts
@@ -1,0 +1,152 @@
+import { CheckResult, BatchResult } from '../report';
+
+export function generateMarkdownReport(results: CheckResult[], targetUrl?: string): string {
+	let blocked = 0;
+	let bypassed = 0;
+	let errors = 0;
+	let other = 0;
+	let detectedWAF = 'Unknown';
+
+	for (const r of results) {
+		if (r.wafType && r.wafType !== 'Unknown') {
+			detectedWAF = r.wafType;
+		}
+		if (r.status === 403 || r.status === '403' || r.status === 'BLOCKED') {
+			blocked++;
+		} else if (r.status === 200 || r.status === '200') {
+			bypassed++;
+		} else if (r.status === 'ERR' || r.status === 500 || r.status === '500') {
+			errors++;
+		} else {
+			other++;
+		}
+	}
+
+	const total = results.length;
+	const protectionScore = total > 0 ? Math.round((blocked / total) * 100) : 100;
+
+	// Calculate per-category stats
+	const categoryMap = new Map<string, { total: number; blocked: number; bypassed: number; errors: number }>();
+
+	for (const r of results) {
+		const entry = categoryMap.get(r.category) || { total: 0, blocked: 0, bypassed: 0, errors: 0 };
+		entry.total++;
+		if (r.status === 403 || r.status === '403' || r.status === 'BLOCKED') {
+			entry.blocked++;
+		} else if (r.status === 200 || r.status === '200') {
+			entry.bypassed++;
+		} else {
+			entry.errors++;
+		}
+		categoryMap.set(r.category, entry);
+	}
+
+	const bypassedItems = results.filter((r) => r.status === 200 || r.status === '200');
+
+	// Status badge
+	const scoreEmoji = protectionScore >= 90 ? '🟢' : protectionScore >= 70 ? '🟡' : '🔴';
+	const statusBadge =
+		protectionScore === 100
+			? '🛡️ **Excellent: 100% Protection**'
+			: protectionScore >= 90
+				? '🛡️ **Good Protection**'
+				: protectionScore >= 70
+					? '⚠️ **Moderate Protection**'
+					: '🚨 **Poor Protection / High Risk**';
+
+	const safeTargetUrl = (targetUrl || 'Target').replace(/`/g, '&#96;');
+
+	const lines: string[] = [
+		`# 🛡️ WAF Checker Audit Report`,
+		``,
+		`> **Target:** \`${safeTargetUrl}\` | **Date:** ${new Date().toUTCString()} | **Detected WAF:** \`${detectedWAF}\``,
+		``,
+		`### ${scoreEmoji} Overall Score: ${protectionScore}% (${statusBadge})`,
+		``,
+		`| Metric | Count | Percentage |`,
+		`| :--- | :--- | :--- |`,
+		`| 🎯 **Total Tests** | \`${total}\` | \`100%\` |`,
+		`| 🛡️ **Blocked (Protected)** | \`${blocked}\` | \`${total > 0 ? Math.round((blocked / total) * 100) : 0}%\` |`,
+		`| ⚠️ **Bypassed (Vulnerable)** | \`${bypassed}\` | \`${total > 0 ? Math.round((bypassed / total) * 100) : 0}%\` |`,
+		`| ❓ **Errors / Other** | \`${errors + other}\` | \`${total > 0 ? Math.round(((errors + other) / total) * 100) : 0}%\` |`,
+		``,
+		`### 📊 Attack Category Breakdown`,
+		``,
+		`| Attack Category | Total | Blocked | Bypassed | Protection Rate |`,
+		`| :--- | :--- | :--- | :--- | :--- |`,
+	];
+
+	for (const [cat, data] of categoryMap.entries()) {
+		const rate = data.total > 0 ? Math.round((data.blocked / data.total) * 100) : 100;
+		const catEmoji = rate >= 90 ? '🟢' : rate >= 70 ? '🟡' : '🔴';
+		lines.push(`| **${cat}** | \`${data.total}\` | \`${data.blocked}\` | \`${data.bypassed}\` | ${catEmoji} \`${rate}%\` |`);
+	}
+
+	lines.push(``);
+
+	if (bypassedItems.length > 0) {
+		lines.push(`### ⚠️ Bypassed Payloads (${bypassedItems.length})`);
+		lines.push(``);
+		lines.push(`<details>`);
+		lines.push(`<summary><b>Click to inspect ${bypassedItems.length} bypassed attack vectors</b></summary>`);
+		lines.push(``);
+		lines.push(`| Category | Method | Status | Time | Bypass Technique | Payload |`);
+		lines.push(`| :--- | :--- | :--- | :--- | :--- | :--- |`);
+
+		for (const item of bypassedItems) {
+			const safePayload = item.payload
+				.replace(/&/g, '&amp;')
+				.replace(/</g, '&lt;')
+				.replace(/>/g, '&gt;')
+				.replace(/"/g, '&quot;')
+				.replace(/'/g, '&#39;')
+				.replace(/\|/g, '&#124;')
+				.replace(/\r?\n/g, ' ');
+			lines.push(
+				`| \`${item.category}\` | \`${item.method}\` | \`${item.status}\` | \`${item.responseTime}ms\` | \`${item.bypassTechnique || 'Standard'}\` | <code>${safePayload}</code> |`,
+			);
+		}
+
+		lines.push(``);
+		lines.push(`</details>`);
+	} else {
+		lines.push(`### ✅ No Bypasses Detected`);
+		lines.push(`All executed attack payloads were successfully blocked by the WAF.`);
+	}
+
+	lines.push(``);
+	lines.push(`---`);
+	lines.push(`*Generated by [WAF-Checker](https://github.com/SecH0us3/waf-checker)*`);
+
+	return lines.join('\n');
+}
+
+export function generateBatchMarkdown(batchResults: BatchResult[]): string {
+	const totalTargets = batchResults.length;
+	const totalBypassed = batchResults.reduce((acc, r) => acc + (r.bypassed || 0), 0);
+	const totalBlocked = batchResults.reduce((acc, r) => acc + (r.blocked || 0), 0);
+	const totalTests = batchResults.reduce((acc, r) => acc + (r.total || 0), 0);
+	const overallRate = totalTests > 0 ? Math.round((totalBlocked / totalTests) * 100) : 100;
+
+	const mdLines = [
+		`# 🛡️ WAF Batch Audit Report`,
+		``,
+		`> **Scanned Targets:** \`${totalTargets}\` | **Total Tests:** \`${totalTests}\` | **Overall Protection:** \`${overallRate}%\``,
+		``,
+		`| Target URL | Status | Total Tests | Blocked | Bypassed | Bypass Rate |`,
+		`| :--- | :--- | :--- | :--- | :--- | :--- |`,
+	];
+
+	for (const r of batchResults) {
+		const statusBadge = !r.success ? '🔴 Failed' : r.bypassed > 0 ? '⚠️ Bypasses' : '🟢 Secure';
+		const safeUrl = (r.url || '').replace(/\|/g, '&#124;').replace(/`/g, '&#96;');
+		mdLines.push(
+			`| \`${safeUrl}\` | ${statusBadge} | \`${r.total || 0}\` | \`${r.blocked || 0}\` | \`${r.bypassed || 0}\` | \`${r.bypassRate || 0}%\` |`,
+		);
+	}
+
+	mdLines.push(``);
+	mdLines.push(`---`);
+	mdLines.push(`*Generated by [WAF-Checker](https://github.com/SecH0us3/waf-checker)*`);
+	return mdLines.join('\n');
+}
