+++ packages/core/src/reports/index.ts
@@ -1,12 +1,8 @@
 export * from './types';
 export * from './sarif';
-export * from './markdown';
-export * from './html';
 
 import { AuditResultItem, calculateAuditStats } from './types';
 import { generateSARIFReport } from './sarif';
-import { generateMarkdownReport } from './markdown';
-import { generateHTMLReport } from './html';
 
 export function generateJSONReport(results: AuditResultItem[], targetUrl?: string): string {
 	const stats = calculateAuditStats(results, targetUrl);
@@ -20,7 +16,7 @@ export function generateJSONReport(results: AuditResultItem[], targetUrl?: strin
 	);
 }
 
-export type ReportFormat = 'sarif' | 'markdown' | 'html' | 'json';
+export type ReportFormat = 'sarif' | 'json';
 
 export function generateReport(
 	format: ReportFormat,
@@ -30,10 +26,6 @@ export function generateReport(
 	switch (format) {
 		case 'sarif':
 			return generateSARIFReport(results, targetUrl);
-		case 'markdown':
-			return generateMarkdownReport(results, targetUrl);
-		case 'html':
-			return generateHTMLReport(results, targetUrl);
 		case 'json':
 		default:
 			return generateJSONReport(results, targetUrl);
