--- packages/core/src/reports/markdown.ts
@@ -1,100 +1,0 @@
-import { AuditResultItem, calculateAuditStats } from './types';
-
-export function generateMarkdownReport(results: AuditResultItem[], targetUrl?: string): string {
-	const stats = calculateAuditStats(results, targetUrl);
-
-	// Calculate per-category stats
-	const categoryMap = new Map<string, { total: number; blocked: number; bypassed: number; errors: number }>();
-
-	for (const r of results) {
-		const entry = categoryMap.get(r.category) || { total: 0, blocked: 0, bypassed: 0, errors: 0 };
-		entry.total++;
-		if (r.status === 403 || r.status === '403' || r.status === 'BLOCKED') {
-			entry.blocked++;
-		} else if (r.status === 200 || r.status === '200') {
-			entry.bypassed++;
-		} else {
-			entry.errors++;
-		}
-		categoryMap.set(r.category, entry);
-	}
-
-	const bypassedItems = results.filter((r) => r.status === 200 || r.status === '200');
-
-	// Status badge
-	const scoreEmoji = stats.protectionScore >= 90 ? '🟢' : stats.protectionScore >= 70 ? '🟡' : '🔴';
-	const statusBadge =
-		stats.protectionScore === 100
-			? '🛡️ **Excellent: 100% Protection**'
-			: stats.protectionScore >= 90
-				? '🛡️ **Good Protection**'
-				: stats.protectionScore >= 70
-					? '⚠️ **Moderate Protection**'
-					: '🚨 **Poor Protection / High Risk**';
-
-	const safeTargetUrl = (stats.targetUrl || 'Target').replace(/`/g, '&#96;');
-
-	const lines: string[] = [
-		`# 🛡️ WAF Checker Audit Report`,
-		``,
-		`> **Target:** \`${safeTargetUrl}\` | **Date:** ${new Date().toUTCString()} | **Detected WAF:** \`${stats.detectedWAF}\``,
-		``,
-		`### ${scoreEmoji} Overall Score: ${stats.protectionScore}% (${statusBadge})`,
-		``,
-		`| Metric | Count | Percentage |`,
-		`| :--- | :--- | :--- |`,
-		`| 🎯 **Total Tests** | \`${stats.total}\` | \`100%\` |`,
-		`| 🛡️ **Blocked (Protected)** | \`${stats.blocked}\` | \`${stats.total > 0 ? Math.round((stats.blocked / stats.total) * 100) : 0}%\` |`,
-		`| ⚠️ **Bypassed (Vulnerable)** | \`${stats.bypassed}\` | \`${stats.total > 0 ? Math.round((stats.bypassed / stats.total) * 100) : 0}%\` |`,
-		`| ❓ **Errors / Other** | \`${stats.errors + stats.other}\` | \`${stats.total > 0 ? Math.round(((stats.errors + stats.other) / stats.total) * 100) : 0}%\` |`,
-		``,
-		`### 📊 Attack Category Breakdown`,
-		``,
-		`| Attack Category | Total | Blocked | Bypassed | Protection Rate |`,
-		`| :--- | :--- | :--- | :--- | :--- |`,
-	];
-
-	for (const [cat, data] of categoryMap.entries()) {
-		const rate = data.total > 0 ? Math.round((data.blocked / data.total) * 100) : 100;
-		const catEmoji = rate >= 90 ? '🟢' : rate >= 70 ? '🟡' : '🔴';
-		lines.push(`| **${cat}** | \`${data.total}\` | \`${data.blocked}\` | \`${data.bypassed}\` | ${catEmoji} \`${rate}%\` |`);
-	}
-
-	lines.push(``);
-
-	if (bypassedItems.length > 0) {
-		lines.push(`### ⚠️ Bypassed Payloads (${bypassedItems.length})`);
-		lines.push(``);
-		lines.push(`<details>`);
-		lines.push(`<summary><b>Click to inspect ${bypassedItems.length} bypassed attack vectors</b></summary>`);
-		lines.push(``);
-		lines.push(`| Category | Method | Status | Time | Bypass Technique | Payload |`);
-		lines.push(`| :--- | :--- | :--- | :--- | :--- | :--- |`);
-
-		for (const item of bypassedItems) {
-			const safePayload = item.payload
-				.replace(/&/g, '&amp;')
-				.replace(/</g, '&lt;')
-				.replace(/>/g, '&gt;')
-				.replace(/"/g, '&quot;')
-				.replace(/'/g, '&#39;')
-				.replace(/\|/g, '&#124;')
-				.replace(/\r?\n/g, ' ');
-			lines.push(
-				`| \`${item.category}\` | \`${item.method}\` | \`${item.status}\` | \`${item.responseTime}ms\` | \`${item.bypassTechnique || 'Standard'}\` | <code>${safePayload}</code> |`,
-			);
-		}
-
-		lines.push(``);
-		lines.push(`</details>`);
-	} else {
-		lines.push(`### ✅ No Bypasses Detected`);
-		lines.push(`All executed attack payloads were successfully blocked by the WAF.`);
-	}
-
-	lines.push(``);
-	lines.push(`---`);
-	lines.push(`*Generated by [WAF-Checker](https://github.com/SecH0us3/waf-checker)*`);
-
-	return lines.join('\n');
-}
